changing the Warning as errors compiler option

+++ testtools/umock_c/tests/umocktypes_charptr_unittests/umocktypes_charptr_unittests.c
@@ -784,7 +784,7 @@ TEST_FUNCTION(umocktypes_charptr_register_types_registers_all_types)
 
     // assert
     ASSERT_ARE_EQUAL(int, 0, result);
-    ASSERT_ARE_EQUAL(int, 2, umocktypes_register_type_call_count);
+    ASSERT_ARE_EQUAL(size_t, 2, umocktypes_register_type_call_count);
     ASSERT_ARE_EQUAL(char_ptr, "char*", umocktypes_register_type_calls[0].type);
     ASSERT_ARE_EQUAL(char_ptr, "const char*", umocktypes_register_type_calls[1].type);
 
@@ -814,7 +814,7 @@ TEST_FUNCTION(when_the_underlying_register_fails_umocktypes_charptr_register_typ
 
         // assert
         ASSERT_ARE_NOT_EQUAL(int, 0, result);
-        ASSERT_ARE_EQUAL(int, i + 1, umocktypes_register_type_call_count);
+        ASSERT_ARE_EQUAL(size_t, i + 1, umocktypes_register_type_call_count);
     }
 }
